requirements-no-pin.txt
hydra-core
loguru
numpy
omegaconf
pandas
pytorch-lightning
pre-commit
pytest
torchvision
